refactor out getRows helper
