+++ src/test/kotlin/dartzee/TestUtils.kt
@@ -153,6 +153,17 @@ fun startThreadMonitoring()
     Thread(threadStackRunnable).start()
 }
 
+fun ScrollTable.getRows(): List<List<Any?>>
+{
+    val result = mutableListOf<List<Any?>>()
+    for (rowIx in 0 until rowCount) {
+        val row = (0 until columnCount).map { getValueAt(rowIx, it) }
+        result.add(row)
+    }
+
+    return result.toList()
+}
+
 /**
  * TODO - improvements for swing-test
  */
